refactor: redesigned insertion sort and swap function and made number types more extensive for hash functions

+++ lib_srcs/basic_utils/includes/hash_funcs.h
@@ -1,6 +1,16 @@
 #ifndef HASH_FUNCS_H
 #define HASH_FUNCS_H
 
-unsigned long djb2Hash(const void *str);
+#include <stddef.h>
+
+/**
+ * @brief
+ *  hash function for strings using the djb2 algorithm.
+ *
+ * @param str   string to hash
+ *
+ * @return hashed index
+ */
+size_t djb2Hash(const void *str);
 
 #endif
