+++ lib_srcs/basic_utils/srcs/hash_funcs.c
@@ -1,9 +1,9 @@
 #include "hash_funcs.h"
 
-unsigned long djb2Hash(const void *str)
+size_t djb2Hash(const void *str)
 {
-    unsigned long hash = 5381;
-    for (int i = 0; ((char *)str)[i] != '\0'; i++) {
+    size_t hash = 5381;
+    for (size_t i = 0; ((char *)str)[i] != '\0'; i++) {
         hash = ((hash << 5) + hash) + ((char *)str)[i]; /* hash*33 + c */
     }
     return hash;
